+++ shell.nix
@@ -1,0 +1,12 @@
+{ pkgs ? import <nixpkgs> { } }:
+
+with pkgs;
+
+mkShell {
+  buildInputs = [
+    stdenv
+    nodejs
+    yarn
+    docker
+  ];
+}